pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k8s-observability-agent"
version = "0.1.0"
description = "AI agent that reverse-engineers K8s platforms from Git repos and designs observability solutions"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.42.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
    "rich>=13.9.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "gitpython>=3.1.0",
    "PyGithub>=2.3",
    "pydantic>=2.8",
    "pathspec>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[project.scripts]
k8s-obs = "agent.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100
